Log out only if api returns 401 status code

+++ src/api/api.ts
@@ -1,7 +1,8 @@
 import axios from 'axios'
 
 declare var localStorage;
-declare function  initializeStorage ();
+declare function initializeStorage ();
+declare function refreshAWSCredentials(callback): any;
 
 let baseUrl = 'https://rw3gl2g6ff.execute-api.us-east-1.amazonaws.com/dev';
 let config = {
@@ -21,8 +22,8 @@ let config = {
 let $http = axios.create(config.axiosConfig)
 
 $http.interceptors.request.use((config) => {
-  if (localStorage.getItem('sessionTokens') || sessionStorage.getItem('sessionTokens')) {
-    const tokenInfo = JSON.parse(localStorage.getItem('sessionTokens') || sessionStorage.getItem('sessionTokens'));
+  if (localStorage.getItem('sessionTokens')) {
+    const tokenInfo = JSON.parse(localStorage.getItem('sessionTokens'));
     config.headers['authorization'] = 'Bearer ' +  tokenInfo['IdToken']['jwtToken'];
   }
   return config;
@@ -33,9 +34,31 @@ $http.interceptors.request.use((config) => {
 $http.interceptors.response.use((response) => {
   return response;
 }, (error) => {
-  localStorage.clear();
-  initializeStorage();
-  window.location.href = '/';
+  const originalRequest = error.config;
+  // if (error.response && error.response.status == 401) {
+  if (typeof error.response === 'undefined') {
+    if (originalRequest._retry) {
+      const firstLogin = localStorage.getItem('firstLogin');
+      localStorage.clear();
+      if (firstLogin === 'no') localStorage.setItem('firstLogin', 'no');
+      initializeStorage();
+      window.location.href = '/';
+    } else {
+      originalRequest._retry = true;
+      let sessionTokens = JSON.parse(localStorage.getItem('sessionTokens'));
+      if (!sessionTokens) {
+        window.location.href = '/';
+      }
+      const idTokenExp = sessionTokens['IdToken']['payload']['exp'];
+      const currentTime = new Date().valueOf();
+      // Check if token is expired, refresh token
+      if (currentTime / 1000 > idTokenExp) {
+        return refreshAWSCredentials(() => {
+          return $http(originalRequest);
+        });
+      }
+    }
+  }
   return Promise.reject(error);
 });
